Turned on whitespace control for Jinja2 block. This behaviour can be turned of with the --no-compact flag. See: http://jinja.pocoo.org/docs/2.10/templates/#whitespace-control
Rationale:
The common use of Jinja2 is to serve web requests in realtime. In this
context, it is reasonable to turn off whitespace control to save some
processing time.
However, j2cli is meant mostly as a batch processing tool. I.e. it is
invoked periodically, the generated output persists for much longer,
and it is more likely to be inspected by a human. Therefore, enabling
whitespace control by default to produce better looking output makes
sense.