setup: Start of a heresy. Renamed to Juiced Jinja2 command-line tool (jj2cli).
The upstream [1] seems to have fallen in disrepair.
In the meantime, this fork has implemented additional functionality and
alternative (and arguably more flexible) approaches to some use cases.
Even at this point, a merge with the upstream is unlikely to happen.
And it will become even more likely with each future commit.

So, starting with this commit, this fork will be developed independently
and under a new name to reflect the additional capabilities.
Perhaps picking-up new ideas from the upstreams, but without any ambition
to be merged with them.

[1] https://github.com/kolypto/j2cli