setup: Start of a heresy. Renamed to Juiced Jinja2 command-line tool (jj2cli).
The upstream [1] appears to have fallen in disrepair.
In the meantime, this fork has implemented additional functionality and
alternative (and arguably more flexible) approaches to some use cases.
At this point, a merge with the upstream is unlikely to happen and it will
become even more likely with each future commit.

Starting with this commit, this fork will be developed independently
and under a new name to reflect the additional capabilities (and
incompatibilities).

[1] https://github.com/kolypto/j2cli