cli options: More robust data spec string parsing. Change of data spec behaviour!
The old code was making an attempt to work with any part of the
data spec missing:
    - data_in:json would interpret the contents of data_in as json.
    - data_in:foo would interpret the contents of data_in as the
      fallback format, and attach them to "foo" in the context.

This resulted in the code being complicated, and also not working
as expected in some cases. For this, it was decided to simplify the
data spec format as <source>|[<source>]:[<format>[:<context_dest]].
Any parts not-required can be ommitted.

The previous examples in the new format can be written as:
    - data_in:json
    - data_in::foo or data_in:?:foo

The order of format and context_dest parts has been reversed, as the
format part will probably be used more often.